+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.4.0] - 2021-08-13
 ### Added
 - Linter's execution action can be set by the `phpCodeSniffer.lintAction` option.
 - Command `phpCodeSniffer.cancelProcessing` to cancel all active processing.